+++ .gitignore
@@ -1,1 +1,1 @@
-DataGenerated/
+*_pycache*
