+++ pyproject.toml
@@ -4,3 +4,6 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools.package-data]
 "files2db" = ["data/file_orga.csv"]
+
+[project.scripts]
+files2db = "files2db.cli:main"
